Split Family.cs into several files.
